move template related components
