+++ src/backoffice/templating/templates/components/index.ts
@@ -1,0 +1,2 @@
+import './template-card/template-card.element';
+import './input-template/input-template.element';
