+++ src/backoffice/templating/templates/components/input-template/input-template.element.ts
@@ -3,7 +3,7 @@ import { UUITextStyles } from '@umbraco-ui/uui-css/lib';
 import { customElement, property, state } from 'lit/decorators.js';
 import { FormControlMixin } from '@umbraco-ui/uui-base/lib/mixins';
 import { UmbTemplateCardElement } from '../template-card/template-card.element';
-import { UmbTemplateRepository } from '../../../templating/templates/repository/template.repository';
+import { UmbTemplateRepository } from '../../repository/template.repository';
 import {
 	UMB_TEMPLATE_PICKER_MODAL,
 	UMB_TEMPLATE_MODAL,
